pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "charmos"
version = "0.4.3"
description = "The Unified System Layer for Agentic Intelligence"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    {name = "ucmind", email = "team@charmos.io"}
]

dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "httpx>=0.25.0",
    "tomlkit>=0.12.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "supabase>=2.0.0",
    "gitpython>=3.1.0",
    "python-dotenv>=1.0.0",
]

[project.urls]
Homepage = "https://charmos.io"
Repository = "https://github.com/CharmAIOS/Charm"

[project.optional-dependencies]
dev = [
    "ruff",
    "pytest",
    "mypy",
    "pre-commit",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
src = ["src", "examples", "tests", "docs"]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "B", "I"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
check_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
addopts = "--tb=short"

[project.scripts]
charm = "charm.cli.app:app"

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
]

[tool.setuptools.package-data]
"charm.contracts" = ["*.json"]
"charm.templates" = ["*.yaml"]